-Making changes suggested by dni. -Adding 3 assertions to remove mypy warnings -Now using require_base_invoice_key instead of require_invoice_key in offer_api.py
